feat: 댓글 예외 처리 추가

+++ src/main/java/app/allstackproject/privideo/service/MyCommentService.java
@@ -1,5 +1,9 @@
 package app.allstackproject.privideo.service;
 
+import static app.allstackproject.privideo.common.response.status.BaseExceptionResponseStatus.COMMENT_NOT_FOUND;
+import static app.allstackproject.privideo.common.response.status.BaseExceptionResponseStatus.COMMENT_UNAUTHORIZED_DELETE;
+
+import app.allstackproject.privideo.common.exception.ApiException;
 import app.allstackproject.privideo.dto.CommentResponse;
 import app.allstackproject.privideo.entity.Comment;
 import app.allstackproject.privideo.repository.CommentRepository;
@@ -19,7 +23,13 @@ public class MyCommentService {
     }
 
     public boolean deleteComment(Long memberId, Long orgId, Long commentId) {
-        //검증 로직 필요
+
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new ApiException(COMMENT_NOT_FOUND));
+
+        if (!comment.getMember().getId().equals(memberId)) {
+            throw new ApiException(COMMENT_UNAUTHORIZED_DELETE);
+        }
 
         commentRepository.deleteById(commentId);
         return true;
